Chore: Cli: Run integration tests in CI
