+++ .eslintignore
@@ -97,7 +97,7 @@ packages/onenote-converter/renderer/pkg/*
 packages/app-cli/app/LinkSelector.js
 packages/app-cli/app/app.js
 packages/app-cli/app/base-command.js
-packages/app-cli/app/cli-integration-tests.js
+packages/app-cli/app/cli-integration-tests.test.js
 packages/app-cli/app/command-apidoc.js
 packages/app-cli/app/command-attach.js
 packages/app-cli/app/command-batch.js
